Final site with concurrency
Sadly , idle time is still higher. So the conclusion is this is not working the way i want it with a less idle time. There is hidden issue that i could not find even though I spent hours on doing it. No need to worry. I am going to improve with more projects and learning . This project gave me a really good knowledge  and understating about Golang  :100: :+1: